DRY MainThread
