+++ core/native/src/main/scala/cats/effect/IOAppPlatform.scala
@@ -16,10 +16,30 @@
 
 package cats.effect
 
-import cats.effect.unsafe.IORuntime
+import cats.effect.unsafe.{IORuntime, UnsafeNonFatal}
+
+import scala.concurrent.ExecutionContext
 
 trait IOAppPlatform extends IOAppCommon with IOAppMultiThreaded {
   this: IOApp =>
+
+  private[effect] def defaultMainThread: ExecutionContext = {
+    new ExecutionContext {
+      def reportFailure(t: Throwable): Unit =
+        t match {
+          case t if UnsafeNonFatal(t) =>
+            IOAppPlatform.this.reportFailure(t).unsafeRunAndForgetWithoutCallback()(runtime)
+
+          case t =>
+            handleTerminalFailure(t)
+        }
+
+      def execute(r: Runnable): Unit =
+        if (!queue.offer(r)) {
+          runtime.blocking.execute(() => queue.put(r))
+        }
+    }
+  }
 
   private[effect] def defaultGlobalRuntime: IORuntime = {
     val (compute, poller, compDown) =
